+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/proposalActionsTypes/index.ts
@@ -1,4 +1,5 @@
 export type { IProposalAction, IProposalActionInputData, IProposalActionInputDataParameter } from './proposalAction';
 export { ProposalActionComponent, type IProposalActionComponentProps } from './proposalActionComponent';
+export type { IProposalActionTokenMint } from './proposalActionTokenMint';
 export { ProposalActionType } from './proposalActionType';
 export type { IProposalActionWithdrawToken } from './proposalActionWithdrawToken';