+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/proposalActionsTypes/proposalActionType.ts
@@ -1,3 +1,4 @@
 export enum ProposalActionType {
     WITHDRAW_TOKEN = 'WITHDRAW_TOKEN',
+    TOKEN_MINT = 'TOKEN_MINT',
 }