+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/proposalActionsUtils.test.tsx
@@ -1,10 +1,12 @@
 import { render, screen } from '@testing-library/react';
+import { generateProposalActionTokenMint } from './actions/generators';
 import { generateProposalAction } from './actions/generators/proposalAction';
 import { generateProposalActionWithdrawToken } from './actions/generators/proposalActionWithdrawToken';
 import { proposalActionsUtils } from './proposalActionsUtils';
 
 jest.mock('./actions', () => ({
     ProposalActionWithdrawToken: () => <div>Mock ProposalActionWithdrawToken</div>,
+    ProposalActionTokenMint: () => <div>Mock ProposalActionTokenMint</div>,
 }));
 
 describe('ProposalActions utils', () => {
@@ -14,6 +16,14 @@ describe('ProposalActions utils', () => {
         const Component = proposalActionsUtils.getActionComponent(action)!;
         render(<Component />);
         expect(screen.getByText('Mock ProposalActionWithdrawToken')).toBeInTheDocument();
+    });
+
+    it('returns ProposalActionTokenMint component for tokenMint action', () => {
+        const action = generateProposalActionTokenMint();
+
+        const Component = proposalActionsUtils.getActionComponent(action)!;
+        render(<Component />);
+        expect(screen.getByText('Mock ProposalActionTokenMint')).toBeInTheDocument();
     });
 
     it('returns null for unknown action type', () => {