+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/proposalActionsUtils.ts
@@ -1,10 +1,17 @@
-import { ProposalActionWithdrawToken } from './actions';
-import { type IProposalAction, type IProposalActionWithdrawToken, ProposalActionType } from './proposalActionsTypes';
+import { ProposalActionTokenMint, ProposalActionWithdrawToken } from './actions';
+import {
+    type IProposalAction,
+    type IProposalActionTokenMint,
+    type IProposalActionWithdrawToken,
+    ProposalActionType,
+} from './proposalActionsTypes';
 
 class ProposalActionsUtils {
     getActionComponent = (action: IProposalAction) => {
         if (this.isWithdrawTokenAction(action)) {
             return () => ProposalActionWithdrawToken({ action });
+        } else if (this.isTokenMintAction(action)) {
+            return () => ProposalActionTokenMint({ action });
         }
 
         return null;
@@ -12,6 +19,10 @@ class ProposalActionsUtils {
 
     isWithdrawTokenAction = (action: Partial<IProposalAction>): action is IProposalActionWithdrawToken => {
         return action.type === ProposalActionType.WITHDRAW_TOKEN;
+    };
+
+    isTokenMintAction = (action: Partial<IProposalAction>): action is IProposalActionTokenMint => {
+        return action.type === ProposalActionType.TOKEN_MINT;
     };
 }